feat: add delete product route and files deletion

+++ src/controllers/products-controller.js
@@ -1,4 +1,6 @@
+const deleteFiles = require("../helpers/deleteFiles")
 const Product = require("../models/Product")
+const path = require("node:path")
 
 module.exports = {
   index: async function (req, res) {
@@ -23,7 +25,8 @@ module.exports = {
       await product.save()
       return res.status(201).json(product)
     } catch (err) {
-      /** @todo handle delete file in case of error */
+      const pathsToDelete = req.files.map(file => file.path)
+      deleteFiles(pathsToDelete)
       return res.status(400).json({ error: err.message })
     }
   },
@@ -37,7 +40,11 @@ module.exports = {
     try {
       const { _id } = req.params
       const product = await Product.findById(_id)
-      /** @todo delete files */
+      if (!product.seller._id.equals(req.user._id)) {
+        return res.status(403).json({ error: "you don't have access to this resource" })
+      }
+      const pathsToDelete = product.images.map(img => path.resolve(process.cwd(), "public", "uploads", "products", img.filename))
+      deleteFiles(pathsToDelete)
       await product.delete()
       return res.status(204).end()
     } catch (err) {
@@ -47,6 +54,6 @@ module.exports = {
 
   /** @todo implement */
   search: async function (req, res) {
-    
+
   }
 }
